fix(hammer): move swipe handling to only over calendar
